requirements.txt
transformers==4.47.0
torch
flask
tensorflow
datasets
safetensors
Pillow
